django-backend/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
Django==5.1.2
django-resized==1.0.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
minio==7.2.10
pillow==11.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pyOpenSSL==24.2.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.1
service-identity==24.2.0
setuptools==75.3.0
sqlparse==0.5.1
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
zope.interface==7.1.1
requests==2.32.3
sqlite-vec==0.1.6
pytz==2025.1
langchain-text-splitters==0.3.7